requirements.txt
cycler==0.10.0
Cython==0.29.14
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
Pillow==6.2.1
pycocotools==2.0.0
pyparsing==2.4.5
python-dateutil==2.8.1
six==1.13.0
torch==1.3.1
torchvision==0.4.2
